Fix lint error in config

+++ config/environment.rb
@@ -1,10 +1,12 @@
 # Load the Rails application.
 require_relative "application"
 
-env_file = File.join(Rails.root, 'config', 'local_env.yml')  
-YAML.load(File.open(env_file)).each do |key, value|
-    ENV[key.to_s] = value  
-end if File.exist?(env_file)
+env_file = Rails.root.join("config/local_env.yml").to_s
+if File.exist?(env_file)
+  YAML.load(File.open(env_file)).each do |key, value|
+    ENV[key.to_s] = value
+  end
+end
 
 # Initialize the Rails application.
 Rails.application.initialize!
